requirements.txt
Pillow==6.2.0
pylint-django==2.0.12
pylint-plugin-utils==0.6
django==2.2.6
django-betterforms==1.2
django-crispy-forms==1.8.0
djangorestframework==3.10.3